Make tints work

+++ shared/world/tint.hh
@@ -1,7 +1,7 @@
 #ifndef A256F04B_6AAF_4B72_A6E4_7667ECF9132A
 #define A256F04B_6AAF_4B72_A6E4_7667ECF9132A
 
-using tint_id_type = std::uint8_t;
+using tint_id_type = std::uint16_t;
 constexpr static tint_id_type TINT_ID_NULL = 0;
 constexpr static tint_id_type TINT_ID_MAX = std::numeric_limits<tint_id_type>::max();
